auth

+++ app/Http/Controllers/QuestionsController.php
@@ -5,16 +5,19 @@ namespace App\Http\Controllers;
 use App\Question;
 use Illuminate\Http\Request;
 use App\Http\Requests\AskQuestionRequest;
+use App\Providers\AuthServiceProvider;
+use Auth;
 
-class QuestionsController extends Controller {
+class QuestionsController extends Controller
+{
 
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index() {
-        
+    public function index()
+    {
         $questions = Question::latest()->paginate(5);
         return view('questions.index', compact('questions'));
     }
@@ -24,29 +27,32 @@ class QuestionsController extends Controller {
      *
      * @return \Illuminate\Http\Response
      */
-    public function create() {
-        $question=new Question();
-        return view('questions.create', compact($question));
+    public function create()
+    {
+        $question = new Question();
+        return view('questions.create', compact('question'));
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
-    public function store(AskQuestionRequest $request) {
-       $request->user()->questions()->create($request->only('title', 'body'));
-       return redirect()->route('questions.index')->with('success','Submited');
+    public function store(AskQuestionRequest $request)
+    {
+        $request->user()->questions()->create($request->only('title', 'body'));
+        return redirect()->route('questions.index')->with('success', 'Submited');
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Question  $question
+     * @param  \App\Question $question
      * @return \Illuminate\Http\Response
      */
-    public function show(Question $question) {
+    public function show(Question $question)
+    {
         $question->increment('views');
         return view('questions.show', compact('question'));
     }
@@ -54,13 +60,16 @@ class QuestionsController extends Controller {
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Question  $question
+     * @param  \App\Question $question
      * @return \Illuminate\Http\Response
      */
-    public function edit(Question $question) {
+    public function edit(Question $question)
+    {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
         return view("questions.edit", compact('question'));
     }
-
     /**
      * Update the specified resource in storage.
      *
@@ -68,20 +77,26 @@ class QuestionsController extends Controller {
      * @param  \App\Question  $question
      * @return \Illuminate\Http\Response
      */
-    public function update(AskQuestionRequest $request, Question $question) {
-        $question->update($request->only('title','body'));
-        return redirect()->route('questions.index')->with('success','Submited');
+    public function update(AskQuestionRequest $request, Question $question)
+    {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Access denied");
+        }
+        $question->update($request->only('title', 'body'));
+        return redirect('/questions')->with('success', "Your question has been updated.");
     }
-
     /**
      * Remove the specified resource from storage.
      *
      * @param  \App\Question  $question
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Question $question) {
+    public function destroy(Question $question)
+    {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Access denied");
+        }
         $question->delete();
-        return redirect()->route('questions.index')->with('success','Deleted');
+        return redirect('/questions')->with('success', "Your question has been deleted.");
     }
-
 }
